change project structure
