+++ example/client-server/go.mod
@@ -2,4 +2,7 @@ module github.com/DistributedClocks/tracing/example/client-server
 
 go 1.14
 
-require github.com/DistributedClocks/tracing v0.0.0-20210122175858-ed6fc3f4eaa9
+require (
+	github.com/DistributedClocks/tracing v0.0.0-20210204210648-537b7afe295b
+	github.com/vmihailenco/msgpack/v5 v5.2.0 // indirect
+)
